fix: preserve gateway's control url
The gateway will only process
"urn:schemas-upnp-org:service:WANIPConnection:1" requests for the
correct URLs.

XML processing is required to extract the correct URL, so a XML library
was added as a dependency. Previously, regex was being used as a
workaround, but it won't work in this case, because there are multiple
"controlURL" tags related to services other than "WANIPConnection:1".

With this patch, rust-igd finally is able to communicate with my router.