implements `remove_port` function
The API is inspired by the gupnp-igd library:
https://developer.gnome.org/gupnp-igd/unstable/GUPnPSimpleIgd.html

And by observations on the UPnP IGD protocol made by Armijn Hemel:
http://www.upnp-hacks.org/igd.html

And by information present in the RFC 6970.

The UPnP procedure uses the name delete, but delete and alike is used as
reserved words in several programming languages, making it a rather
unpopular and avoided name in libraries in general. The gupnp-igd
library doesn't use the name delete. The MongoDB drivers use the name
remove. Rust standard libraries (see HashSet, for instance) also favors
the name remove. Here, we also use the name remove for the user-exposed
API.

Lastly, this API should support IGD 1.0 and IGD 2.0. But for now, only
IGD 1.0 is supported.

Fixes #1