Update parsing.rs
In the code, specifically parsing.rs on line 53 a check is made for urn:schemas-upnp-org:service:WANIPConnection:1, the closest thing I have in my xml is urn:schemas-upnp-org:service:WANPPPConnection1 and changing it to that gets the example to run. I don't really have much experience with networking but all routers in South Africa use PPP as far as I know, perhaps I could open a pull request to check for either of the above and use the appropriate one?