Rewrite aio search without impl Future
The concrete implementation of Future resulted in hard to read code,
with loops and pattern matching and passing status objects around.

Since we do not introduce new polling mechanisms here, but instead rely
on the implementations of hyper and tokio, it is sufficient to work with
async/.await instead of providing our own implementation of Future.

The resulting code is much shorter and easier to understand.